refactor: voice.rs phase 1 — extract doctor + status (UPI 050)
voice.rs at 8940 lines housed 24 independent render surfaces. Phase 1
of UPI 050 converts src/voice.rs → src/voice/mod.rs and extracts the
two largest renderers (doctor + status, the most complex by far) into
their own sub-modules. Remaining 14 renderers stay in voice/mod.rs
pending phase 2.

Extracted:
- src/voice/doctor.rs (555 lines): render_doctor + private helpers —
  format_churn_row, render_shape_kv, render_cost_delta,
  render_reason_line, format_recommendation_row,
  render_doctor_check_section, check_icon_style, is_synth_pointer.
- src/voice/status.rs (436 lines): render_status + private helpers —
  render_status_daemon, render_status_interactive, render_summary_line,
  render_subvolume_table, render_thread_status, format_count_with_age,
  render_advisories, render_redundancy_advisories, render_drive_summary,
  render_last_run.

Cross-renderer helpers stay in voice/mod.rs as pub(super) so sub-modules
can access them via super::: humanize_duration, exposure_label,
format_status_table, strip_ansi_len, color_and_pad, color_exposure_str,
color_health_str, color_result, pluralize, classify_verify_checks,
aggregate_drive_info, unmounted_drive_label, format_drive_age_label,
status_severity, append_suggestion, SuggestionContext, suggest_next_action.

Public surface unchanged via re-exports in voice/mod.rs:
  pub use doctor::render_doctor;
  pub use status::render_status;
Zero changes at caller sites — all 23 importers continue to resolve
`crate::voice::render_status(...)` etc.

Voice Contract regression gate:
- pre-split: 44 tests pass
- post-split: 44 tests pass (identical golden strings)

cargo clippy --all-targets -- -D warnings + cargo test (1435 passed) +
cargo build --release all green.

Phase 2 (remaining 14 renderers — backup, plan, history, verify,
calibrate, init, sentinel, emergency, drives, default_status, etc.)
is a mechanical follow-up using the same pattern. Held for now per
the "skip ceremony" mandate; the locality win drops off fast after
the big two.
